requirements.txt
humanfriendly  # purely optional
git+git://github.com/Lucas-C/pyfpdf.git@master
pikepdf
pillow
pyduktape
tqdm  # purely optional
